We now no longer require an admin user be able to take a tool out-of-service.
We also allow any known user to be able to take the tool out of
service, and log the state change.

We also no longer require the tool to be in-service to return the
permission level of the requesting user. This allows admin users to
be able to use a tool even if it's out of service (so they can fix
the problem). It's now up to the acnode to check that the tool is
in-service, since it can then grant/deny access based on the
user's admin rights.